refactor : ComposeView 상단 Information 담는 View 별도로 분리
